+++ nba-sim-core/src/main/java/ru/plifis/nbasim/repository/SkillsetRepository.java
@@ -3,5 +3,5 @@ package ru.plifis.nbasim.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.plifis.nbasim.model.SkillSetEntity;
 
-public interface SkillsetRepository extends CrudRepository<SkillSetEntity, Long> {
+public interface SkillsetRepository extends CrudRepository<SkillSetEntity, Integer> {
 }
