+++ nba-sim-core/src/main/java/ru/plifis/nbasim/repository/TeamRepository.java
@@ -7,9 +7,9 @@ import ru.plifis.nbasim.model.TeamEntity;
 import java.util.Collection;
 
 @Repository
-public interface TeamRepository extends CrudRepository<TeamEntity, Long> {
+public interface TeamRepository extends CrudRepository<TeamEntity, Integer> {
 
-    TeamEntity findTeamEntityById(Long id);
+    TeamEntity findTeamEntityById(Integer id);
 
     Collection<TeamEntity> findAll();
 }
